refactor(ui): reuse shared Button component and normalize radius sizes
